持续化报告

+++ AutoTest/common/install.py
@@ -8,6 +8,7 @@ from ..common.gold import SmokeThread
 from AutoUI.models import autoui, auto_uirecord
 import time, datetime
 import logging
+from ..utils.keycloak.keycloakadmin import KeycloakAdmin
 
 logger = logging.getLogger(__name__)  # 这里使用 __name__ 动态搜索定义的 logger 配置
 
@@ -92,11 +93,12 @@ class InstallThread(threading.Thread):
             if self.obj.installstatus is True:
                 deldata(self.obj.server, self.id, self.pwd)
             if self.obj.version:
-                self.localpath = '/files/History_version/{0}/{1}.zip'.format(self.obj.version, self.obj.version)
-                if not os.path.exists(self.localpath):
-                    self.obj.type = 2
-                    self.obj.save()
-                    self.downFile(version=self.obj.version)
+                if self.Flag is True:
+                    self.localpath = '/files/History_version/{0}/{1}.zip'.format(self.obj.version, self.obj.version)
+                    if not os.path.exists(self.localpath):
+                        self.obj.type = 2
+                        self.obj.save()
+                        self.downFile(version=self.obj.version)
             else:
                 self.obj.type = 2
                 self.obj.save()
@@ -104,35 +106,38 @@ class InstallThread(threading.Thread):
             self.obj.type = 3
             self.obj.save()
             if not os.path.exists("/home/biomind/{}".format(self.obj.version)):
-                logger.info("Installation{0}：上传最新安装包{1}.zip".format(self.id,self.obj.version))
-                self.ssh.upload(self.localpath, "/home/biomind/QaInstall.zip")
-                self.ssh.cmd("sshpass -p {0} sudo rm -rf {1}/".format(self.pwd,self.obj.version))
-                logger.info("Installation{0}：解压安装包 QaInstall.zip".format(self.id))
-                self.ssh.cmd("unzip {}".format("QaInstall.zip"))
+                if self.Flag is True:
+                    logger.info("Installation{0}：上传最新安装包{1}.zip".format(self.id,self.obj.version))
+                    self.ssh.upload(self.localpath, "/home/biomind/QaInstall.zip")
+                    self.ssh.cmd("sshpass -p {0} sudo rm -rf {1}/".format(self.pwd,self.obj.version))
+                    logger.info("Installation{0}：解压安装包 QaInstall.zip".format(self.id))
+                    self.ssh.cmd("unzip {}".format("QaInstall.zip"))
             logger.info("Installation{}：停止服务".format(self.id))
             self.ssh.cmd("sshpass -p {} biomind stop;".format(self.pwd))
             logger.info("Installation{}：安装最新版本".format(self.id))
             self.ssh.cmd("cd {0};sshpass -p {1} bash setup_engine.sh;".format(self.obj.version, self.pwd))
             time.sleep(5)
             try:
-                logger.info("Installation{}：更新orthanc文件".format(self.id))
-                self.ssh.upload("/files1/classifier/orthanc.json".format(settings.BASE_DIR),
+                logger.info("Installation{}：更新 orthanc 文件".format(self.id))
+                self.ssh.upload("/files1/classifier/orthanc.json",
                                 "/home/biomind/.biomind/var/biomind/orthanc/orthanc.json")
 
                 logger.info("Installation{}：更新classification_votes文件".format(self.id))
-                self.ssh.upload("/files1/classifier/classification_votes.json".format(settings.BASE_DIR),
+                self.ssh.upload("/files1/classifier/classification_votes.json",
                                 "/home/biomind/.biomind/var/biomind/cache/series_classifier/classification_votes.json")
 
                 logger.info("Installation{}：更新predefined_classifier文件".format(self.id))
-                self.ssh.upload("/files1/classifier/predefined_classifier.json".format(settings.BASE_DIR),
+                self.ssh.upload("/files1/classifier/predefined_classifier.json",
                                 "/home/biomind/.biomind/var/biomind/cache/series_classifier/predefined_classifier.json")
 
                 logger.info("Installation{}：更新special_classifier文件".format(self.id))
-                self.ssh.upload("/files1/classifier/special_classifier.json".format(settings.BASE_DIR),
+                self.ssh.upload("/files1/classifier/special_classifier.json",
                                 "/home/biomind/.biomind/var/biomind/cache/series_classifier/special_classifier.json")
             except Exception as e:
                 logger.error("Installation{0}：更新json文件失败----失败原因：{1}".format(self.id, e))
             self.restart()
+            logger.info("Installation{}：sheep 200 秒".format(self.id))
+            time.sleep(200)
             self.goldsmoke()
         except Exception as e:
             self.obj.status = False
@@ -141,33 +146,35 @@ class InstallThread(threading.Thread):
 
     def restart(self):
         try:
-            logger.info("Installation{}：配置 configure".format(self.id))
-            self.ssh.configure(self.obj.server, str(self.obj.Host.protocol))
-            self.obj.type = 4
-            self.obj.save()
-            logger.info("Installation{}：重启服务".format(self.id))
-            self.ssh.cmd("sshpass -p {} biomind start prod master;".format(self.pwd))
-            self.ssh.close()
-            time.sleep(200)
-            # try:
-            #     groups = KeycloakAdmin.get_groups()
-            #     KeycloakAdmin.create_user(
-            #     'test', 'Asd@123456', [
-            #         x['id'] for x in groups])  # 获得角色(分组)ID
-
-            # except Exception as e:
-            #     logging.error('Failed to create User: %s!', e)
+            if self.Flag is True:
+                logger.info("Installation{}：配置 configure".format(self.id))
+                self.ssh.configure(self.obj.server, str(self.obj.Host.protocol))
+                self.obj.type = 4
+                self.obj.save()
+                logger.info("Installation{}：重启服务".format(self.id))
+                self.ssh.cmd("sshpass -p {} biomind start prod master;".format(self.pwd))
+                self.ssh.close()
         except Exception as e:
             self.obj.status = False
             self.obj.save()
             logger.error("Installation{0}：重启服务失败{1}".format(self.id,e))
+
+    def createUser(self):
+        try:
+            if self.Flag is True:
+                groups = KeycloakAdmin.get_groups()
+                KeycloakAdmin.create_user(
+                'test', 'Asd@123456', [
+                    x['id'] for x in groups])  # 获得角色(分组)ID
+        except Exception as e:
+            logging.error('Failed to create User: %s!', e)
 
     def goldsmoke(self):
         try:
             data = {"version": self.obj.version,
                     "diseases": "19,44,31,32,21,33,23,24,20,30,22,25,26",
                     "status": True,
-                    "hostid": self.obj.Host_id,
+                    "Host_id": self.obj.Host_id,
                     "thread": 1,
                     "count": 127
                     }
